requirements.txt
Flask==1.0.2
Flask-Argon2==0.1.2.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-Principal==0.4.0
Flask-Login==0.4.1
Flask-WTF==0.14.2
mysqlclient==1.3.13